+++ src/proyectoCoffeGame/controllers/consola/nuevoConsola.java
@@ -16,7 +16,7 @@ public class nuevoConsola extends JDialog {
         super(parent, "Ingresar Datos Nuevo Computador - COFFE GAMER", true);
         setDefaultCloseOperation(DISPOSE_ON_CLOSE);
         setResizable(false);
-    
+
         JPanel panel = new JPanel(new GridBagLayout());
         panel.setBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10));
         panel.setBackground(Color.WHITE);
@@ -80,27 +80,27 @@ public class nuevoConsola extends JDialog {
         pack();
         setLocationRelativeTo(parent);
     }
-    
+
     private void onAceptar(ActionEvent e) {
         // Obtener los valores ingresados en los campos de texto
         String numSerieC = numSerie.getText();
         String modelC = modelo.getText();
         String precioHS = precioH.getText();
-    
+
         String patronNumerico = "\\d+\\.?\\d*"; // Patrón para validar números
-    
+
         if (numSerieC.isEmpty() || modelC.isEmpty() || precioHS.isEmpty()) {
             JOptionPane.showMessageDialog(null, "Campos no llenos");
         } else if (!precioHS.matches(patronNumerico)) {
             JOptionPane.showMessageDialog(null, "Precio no válido, ingresa un número válido");
         } else {
             try {
-                float precioHC = Float.parseFloat(precioHS);
-                consolaModel compConsola=new consolaModel();
+                int precioHC = Integer.parseInt(precioHS);
+                consolaModel compConsola = new consolaModel();
                 compConsola.setNumSerie(numSerieC);
                 compConsola.setModelo(modelC);
                 compConsola.setPrecioH(precioHC);
-    
+
                 // Supongo que nuevoCliente se inicializa en algún lugar antes de este código
                 consolaController.insertarConsola(compConsola);
                 dispose();
@@ -109,6 +109,5 @@ public class nuevoConsola extends JDialog {
             }
         }
     }
-    
-    
+
 }
